+++ android-sdk-ui/src/main/java/com/appboy/ui/inappmessage/listeners/IWebViewClientStateListener.java
@@ -1,7 +1,0 @@
-package com.appboy.ui.inappmessage.listeners;
-
-/**
- * @deprecated Please use {@link com.braze.ui.inappmessage.listeners.IWebViewClientStateListener} instead. Deprecated since 6/10/21
- */
-@Deprecated
-public interface IWebViewClientStateListener extends com.braze.ui.inappmessage.listeners.IWebViewClientStateListener {}