+++ android-sdk-ui/src/main/java/com/braze/ui/inappmessage/IInAppMessageAnimationFactory.kt
@@ -1,7 +1,7 @@
 package com.braze.ui.inappmessage
 
-import com.braze.models.inappmessage.IInAppMessage
 import android.view.animation.Animation
+import com.braze.models.inappmessage.IInAppMessage
 
 interface IInAppMessageAnimationFactory {
     /**
@@ -9,12 +9,12 @@ interface IInAppMessageAnimationFactory {
      * @return animation that will be applied to the in-app message view using
      * [android.view.View.setAnimation]
      */
-    fun getOpeningAnimation(inAppMessage: IInAppMessage): Animation
+    fun getOpeningAnimation(inAppMessage: IInAppMessage): Animation?
 
     /**
      * This method returns the animation that will be used to animate the message as it exits the screen.
      * @return animation that will be applied to the in-app message view using
      * [android.view.View.setAnimation]
      */
-    fun getClosingAnimation(inAppMessage: IInAppMessage): Animation
+    fun getClosingAnimation(inAppMessage: IInAppMessage): Animation?
 }