+++ android-sdk-ui/src/main/java/com/braze/ui/inappmessage/listeners/IWebViewClientStateListener.kt
@@ -1,0 +1,8 @@
+package com.braze.ui.inappmessage.listeners
+
+fun interface IWebViewClientStateListener {
+    /**
+     * Fired when [android.webkit.WebViewClient.onPageFinished] has been called.
+     */
+    fun onPageFinished()
+}