+++ android-sdk-ui/src/main/java/com/braze/ui/inappmessage/views/InAppMessageHtmlBaseView.java
@@ -194,6 +194,7 @@ public abstract class InAppMessageHtmlBaseView extends RelativeLayout implements
       InAppMessageViewUtils.closeInAppMessageOnKeycodeBack();
       return true;
     }
+    ViewUtils.setFocusableInTouchModeAndRequestFocus(mMessageWebView);
     return super.onKeyDown(keyCode, event);
   }